update schemas and tests
