+++ backend/tests/unit/test_story_service.py
@@ -20,7 +20,6 @@ from app.db.notification import Notification
 from app.db.tag import Tag
 from app.models.comment import CommentCreateRequest
 from app.models.story import MediaUploadRequest, StoryCreateRequest, StoryResponse, StoryUpdateRequest
-from app.services.admin_service import remove_story_as_admin
 from app.services.story_service import (
     create_comment_for_story,
     create_story_with_location,
@@ -1783,45 +1782,6 @@ class TestAiTagPersistenceHelpers:
 
         with pytest.raises(HTTPException) as exc_info:
             await apply_ai_tags_to_story(db, uuid.uuid4(), ["bogazici"])
-
-        assert exc_info.value.status_code == 404
-        assert exc_info.value.detail == "Story not found"
-        db.commit.assert_not_awaited()
-
-
-@pytest.mark.asyncio
-class TestAdminRemoveStoryService:
-    async def test_soft_deletes_story_and_marks_pending_reports_removed(self):
-        story_id = uuid.uuid4()
-        admin_id = uuid.uuid4()
-        story = _make_story(id=story_id, deleted_at=None, deleted_by=None)
-        current_user = SimpleNamespace(id=admin_id)
-
-        pending_report_1 = SimpleNamespace(status=ReportStatus.PENDING)
-        pending_report_2 = SimpleNamespace(status=ReportStatus.PENDING)
-
-        db = AsyncMock()
-        db.add = MagicMock()
-        db.execute.side_effect = [
-            SimpleNamespace(scalar_one_or_none=lambda: story),
-            SimpleNamespace(scalars=lambda: SimpleNamespace(all=lambda: [pending_report_1, pending_report_2])),
-        ]
-
-        await remove_story_as_admin(db, story_id, current_user)
-
-        assert story.deleted_at is not None
-        assert story.deleted_by == admin_id
-        assert pending_report_1.status == ReportStatus.REMOVED
-        assert pending_report_2.status == ReportStatus.REMOVED
-        db.commit.assert_awaited_once()
-        assert db.add.call_count == 3
-
-    async def test_remove_story_raises_404_when_story_missing(self):
-        db = AsyncMock()
-        db.execute.return_value.scalar_one_or_none = lambda: None
-
-        with pytest.raises(HTTPException) as exc_info:
-            await remove_story_as_admin(db, uuid.uuid4(), SimpleNamespace(id=uuid.uuid4()))
 
         assert exc_info.value.status_code == 404
         assert exc_info.value.detail == "Story not found"
